# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for E-Document Management System",
        "title": "E-Document API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@edocument.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api",
    "paths": {
...
        "/v1/upload/download/folder/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Downloads all files in a folder (including subfolders) as a ZIP archive",
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Download a folder as ZIP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/upload/download/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Downloads a file by attachment ID with original filename",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Download a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attachment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/upload/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns information about the TUS upload service including supported version and extensions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Get upload service info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_app_upload.UploadInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "e-document-backend_internal_util.ErrorCode": {
            "type": "string",
            "enum": [
                "UNAUTHORIZED",
                "INCORRECT_PASSWORD",
                "INVALID_CREDENTIALS",
                "TOKEN_EXPIRED",
                "INVALID_TOKEN",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "MISSING_REQUIRED_FIELD",
                "INVALID_INPUT",
                "INTERNAL_SERVER_ERROR",
                "DATABASE_ERROR",
                "CONFIG_NOT_SET",
                "USER_NOT_FOUND",
                "USER_ALREADY_EXISTS",
                "EMAIL_ALREADY_EXISTS",
                "ROLE_NOT_FOUND",
                "ROLE_ALREADY_EXISTS"
            ],
            "x-enum-varnames": [
                "UNAUTHORIZED",
                "INCORRECT_PASSWORD",
                "INVALID_CREDENTIALS",
                "TOKEN_EXPIRED",
                "INVALID_TOKEN",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "MISSING_REQUIRED_FIELD",
                "INVALID_INPUT",
                "INTERNAL_SERVER_ERROR",
                "DATABASE_ERROR",
                "CONFIG_NOT_SET",
                "USER_NOT_FOUND",
                "USER_ALREADY_EXISTS",
                "EMAIL_ALREADY_EXISTS",
                "ROLE_NOT_FOUND",
                "ROLE_ALREADY_EXISTS"
            ]
        },
...
        "e-document-backend_internal_util.PaginationInfo": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "itemsPerPage": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "e-document-backend_internal_util.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "$ref": "#/definitions/e-document-backend_internal_util.ErrorCode"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/e-document-backend_internal_util.PaginationInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_app_upload.UploadInfoResponse": {
            "type": "object",
            "properties": {
                "extensions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "creation",
                        "creation-defer-length",
                        "termination",
                        "concatenation"
                    ]
                },
                "max_size": {
                    "type": "integer",
                    "example": 0
                },
                "tus_version": {
                    "type": "string",
                    "example": "1.0.0"
                },
                "upload_path": {
                    "type": "string",
                    "example": "/api/v1/upload/files"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}